-Further defined AJAX delete for authors and books. Both working 100%

+++ src/main/webapp/authorAJAX.js
@@ -3,21 +3,21 @@
  */
 
 (function ($, window, document) {
-    
+
     $(function () {
-        
-        var $btnDeleteAuthors = $('#deleteAuthor');        
-        var baseUrl = 'bc';  
-            
-        function rerenderList(){
-            $('.checkedAuthors:checked').parent().parent().remove(); 
-            
+
+        var $btnDeleteAuthors = $('#deleteAuthor');
+        var baseUrl = 'ac';
+
+        function rerenderList() {
+            $('.checkedAuthors:checked').parent().parent().remove();
+
             var $tr = $('#authorTable > tbody > tr').not(':first');
             $tr.each(function (index, tr) {
-                if(index%2!==0){
-                    $(tr).css('background-color','#99e699');
-                }else{
-                    $(tr).css('background-color','transparent');
+                if (index % 2 !== 0) {
+                    $(tr).css('background-color', '#99e699');
+                } else {
+                    $(tr).css('background-color', 'transparent');
                 }
             });
         }
@@ -25,15 +25,19 @@
         function handleError(xhr, status, error) {
             alert("Sorry, there was a problem: " + error);
         }
-        
+
         $btnDeleteAuthors.click(function () {
-            deleteSelectedAuthors();
-            return false;
+            if(!confirm("Selected authors will be deleted. Proceed?")){
+                return false;
+            }else{
+                deleteSelectedAuthors();
+                return false;
+            }
         });
-        
+
         function deleteSelectedAuthors() {
             console.log('deleting selected Authors');
-            
+
             $.ajax({
                 type: 'POST',
                 contentType: 'application/json',
@@ -41,24 +45,24 @@
                 dataType: "json",
                 data: authorIdsToJSON()
             })
-            .done(function () {
-                rerenderList();           
-            })
-            .fail(function ( jqXHR, textStatus, errorThrown ) {
-                alert("Authors could not be deleted due to: " + errorThrown);
-            });
+                    .done(function () {
+                        rerenderList();
+                    })
+                    .fail(function (jqXHR, textStatus, errorThrown) {
+                        alert("Authors could not be deleted due to: " + errorThrown);
+                    });
         }
-        
+
     });
-    
-    function authorIdsToJSON(){
+
+    function authorIdsToJSON() {
         var $authorInputs = $('.checkedAuthors:checked');
         var authorIds = [];
-        for(var i = 0 ; i < $authorInputs.length ; i++){
-           authorIds.push($authorInputs[i].value);     
+        for (var i = 0; i < $authorInputs.length; i++) {
+            authorIds.push($authorInputs[i].value);
         }
-        return JSON.stringify({"authorIds":authorIds});
-        
+        return JSON.stringify({"authorIds": authorIds});
+
     }
 
 }(window.jQuery, window, document));
